requirements.txt
fastapi[all]
openai
typing-extensions
scipy
numpy
nvidia-riva-client
Deepgram-sdk
starlette
